+++ src/Selenium2Library/keywords/browsermanagement.py
@@ -1,13 +1,25 @@
+#Jari
 import os, signal
 import robot
-from robot.errors import DataError
 from robot.api import logger
+#Jari
+
+import os.path
+import time
+import types
+
+from robot.errors import DataError
+from robot.utils import secs_to_timestr, timestr_to_secs
+
 from selenium import webdriver
-from Selenium2Library import webdrivermonkeypatches
+from selenium.webdriver.remote.webdriver import WebDriver as RemoteWebDriver
+from selenium.common.exceptions import NoSuchWindowException
+
 from Selenium2Library.utils import BrowserCache
 from Selenium2Library.locators import WindowManager
-from keywordgroup import KeywordGroup
-from selenium.common.exceptions import NoSuchWindowException
+
+from .keywordgroup import KeywordGroup
+
 
 from selenium.webdriver.remote.errorhandler import ErrorHandler
 from selenium.webdriver.remote.remote_connection import RemoteConnection
@@ -62,12 +74,13 @@ class _ReusableDriver(webdriver.Remote):
         self.capabilities = response['value']
         self.w3c = "specificationLevel" in self.capabilities
 
-class _BrowserManagementKeywords(KeywordGroup):
+
+class BrowserManagementKeywords(KeywordGroup):
 
     def __init__(self):
         self._cache = BrowserCache()
         self._window_manager = WindowManager()
-        self._speed_in_secs = float(0)
+        self._speed_in_secs = 0.0
         self._timeout_in_secs = float(5)
         self._implicit_wait_in_secs = float(0)
 
@@ -421,7 +434,7 @@ class _BrowserManagementKeywords(KeywordGroup):
         return self._current_browser().set_window_size(width, height)
 
     def get_window_position(self):
-        """Returns current window position as `x` then `y`.
+        """Returns current window position as `x` then `y` (relative to the left and top of the screen).
 
         Example:
         | ${x} | ${y}= | Get Window Position |
@@ -430,13 +443,13 @@ class _BrowserManagementKeywords(KeywordGroup):
         return position['x'], position['y']
 
     def set_window_position(self, x, y):
-        """Sets the position `x` and `y` of the current window to the specified values.
+        """Sets the position x and y of the current window (relative to the left and top of the screen) to the specified values.
 
         Example:
-        | Set Window Size | ${1000} | ${0}       |
-        | ${x} | ${y}= | Get Window Position |
-        | Should Be Equal | ${x}      | ${1000}   |
-        | Should Be Equal | ${y}      | ${0}      |
+        | Set Window Position | ${8}    | ${10}               |
+        | ${x}                | ${y}=   | Get Window Position |
+        | Should Be Equal     | ${x}    | ${8}                |
+        | Should Be Equal     | ${y}    | ${10}               |
         """
         return self._current_browser().set_window_position(x, y)
 
@@ -485,7 +498,7 @@ class _BrowserManagementKeywords(KeywordGroup):
         | Select Window |  | | # Chooses the main window again |
         """
         try:
-            return self._current_browser().get_current_window_handle()
+            return self._current_browser().current_window_handle
         except NoSuchWindowException:
             pass
         finally:
@@ -493,7 +506,7 @@ class _BrowserManagementKeywords(KeywordGroup):
 
     def list_windows(self):
         """Return all current window handles as a list"""
-        return self._current_browser().get_window_handles()
+        return self._current_browser().window_handles
 
     def unselect_frame(self):
         """Sets the top frame as the current frame."""
@@ -503,15 +516,22 @@ class _BrowserManagementKeywords(KeywordGroup):
 
     def get_location(self):
         """Returns the current location."""
-        return self._current_browser().get_current_url()
+        return self._current_browser().current_url
+
+    def get_locations(self):
+        """Returns and logs current locations of all windows known to the browser."""
+        return self._log_list(
+            [window_info[4] for window_info in
+             self._window_manager._get_window_infos(self._current_browser())]
+        )
 
     def get_source(self):
         """Returns the entire html source of the current page or frame."""
-        return self._current_browser().get_page_source()
+        return self._current_browser().page_source
 
     def get_title(self):
         """Returns title of current page."""
-        return self._current_browser().get_title()
+        return self._current_browser().title
 
     def location_should_be(self, url):
         """Verifies that current URL is exactly `url`."""
@@ -580,34 +600,42 @@ class _BrowserManagementKeywords(KeywordGroup):
         """Gets the delay in seconds that is waited after each Selenium command.
 
         See `Set Selenium Speed` for an explanation."""
-        return robot.utils.secs_to_timestr(self._speed_in_secs)
+        return secs_to_timestr(self._speed_in_secs)
 
     def get_selenium_timeout(self):
         """Gets the timeout in seconds that is used by various keywords.
 
         See `Set Selenium Timeout` for an explanation."""
-        return robot.utils.secs_to_timestr(self._timeout_in_secs)
+        return secs_to_timestr(self._timeout_in_secs)
 
     def get_selenium_implicit_wait(self):
         """Gets the wait in seconds that is waited by Selenium.
 
         See `Set Selenium Implicit Wait` for an explanation."""
-        return robot.utils.secs_to_timestr(self._implicit_wait_in_secs)
+        return secs_to_timestr(self._implicit_wait_in_secs)
 
     def set_selenium_speed(self, seconds):
         """Sets the delay in seconds that is waited after each Selenium command.
 
         This is useful mainly in slowing down the test execution to be able to
         view the execution. `seconds` may be given in Robot Framework time
-        format. Returns the previous speed value.
+        format. Returns the previous speed value in seconds.
+
+        One keyword may execute one or many Selenium commands and therefore
+        one keyword may slow down more than the ``seconds`` argument defines.
+        Example if delay is set to 1 second and because `Click Element`
+        executes two Selenium commands, then the total delay will be 2 seconds.
+        But because `Page Should Contain Element` executes only one selenium
+        command, then the total delay will be 1 second.
 
         Example:
         | Set Selenium Speed | .5 seconds |
         """
-        old_speed = self.get_selenium_speed()
-        self._speed_in_secs = robot.utils.timestr_to_secs(seconds)
+        old_speed = self._speed_in_secs
+        self._speed_in_secs = timestr_to_secs(seconds)
         for browser in self._cache.browsers:
-            browser.set_speed(self._speed_in_secs)
+            browser._speed = self._speed_in_secs
+            self._monkey_patch_speed(browser)
         return old_speed
 
     def set_selenium_timeout(self, seconds):
@@ -628,7 +656,7 @@ class _BrowserManagementKeywords(KeywordGroup):
         | Set Selenium Timeout | ${orig timeout} |
         """
         old_timeout = self.get_selenium_timeout()
-        self._timeout_in_secs = robot.utils.timestr_to_secs(seconds)
+        self._timeout_in_secs = timestr_to_secs(seconds)
         for browser in self._cache.get_open_browsers():
             browser.set_script_timeout(self._timeout_in_secs)
         return old_timeout
@@ -647,7 +675,7 @@ class _BrowserManagementKeywords(KeywordGroup):
         | Set Selenium Implicit Wait | ${orig wait} |
         """
         old_wait = self.get_selenium_implicit_wait()
-        self._implicit_wait_in_secs = robot.utils.timestr_to_secs(seconds)
+        self._implicit_wait_in_secs = timestr_to_secs(seconds)
         for browser in self._cache.get_open_browsers():
             browser.implicitly_wait(self._implicit_wait_in_secs)
         return old_wait
@@ -665,7 +693,7 @@ class _BrowserManagementKeywords(KeywordGroup):
 
         See also `Set Selenium Implicit Wait`.
         """
-        implicit_wait_in_secs = robot.utils.timestr_to_secs(seconds)
+        implicit_wait_in_secs = timestr_to_secs(seconds)
         self._current_browser().implicitly_wait(implicit_wait_in_secs)
 
     # Private
@@ -687,7 +715,6 @@ class _BrowserManagementKeywords(KeywordGroup):
             raise ValueError(browser_name + " is not a supported browser.")
 
         browser = creation_func(remote, desired_capabilities, profile_dir)
-        browser.set_speed(self._speed_in_secs)
         browser.set_script_timeout(self._timeout_in_secs)
         browser.implicitly_wait(self._implicit_wait_in_secs)
 
@@ -785,3 +812,17 @@ class _BrowserManagementKeywords(KeywordGroup):
             desired_capabilities[key.strip()] = value.strip()
 
         return desired_capabilities
+
+    def _get_speed(self, browser):
+        return browser._speed if hasattr(browser, '_speed') else 0.0
+
+    def _monkey_patch_speed(self, browser):
+        def execute(self, driver_command, params=None):
+            result = self._base_execute(driver_command, params)
+            speed = self._speed if hasattr(self, '_speed') else 0.0
+            if speed > 0:
+                time.sleep(speed)
+            return result
+        if not hasattr(browser, '_base_execute'):
+            browser._base_execute = browser.execute
+            browser.execute = types.MethodType(execute, browser)