+++ src/Selenium2Library/keywords/javascript.py
@@ -1,8 +1,9 @@
 import os
-from keywordgroup import KeywordGroup
+
+from .keywordgroup import KeywordGroup
 
 
-class _JavaScriptKeywords(KeywordGroup):
+class JavaScriptKeywords(KeywordGroup):
 
     # Public