+++ src/Selenium2Library/keywords/waiting.py
@@ -1,15 +1,18 @@
 import time
-import robot
-from keywordgroup import KeywordGroup
 
-class _WaitingKeywords(KeywordGroup):
+from robot.utils import secs_to_timestr, timestr_to_secs
+
+from .keywordgroup import KeywordGroup
+
+
+class WaitingKeywords(KeywordGroup):
 
     # Public
 
     def wait_for_condition(self, condition, timeout=None, error=None):
         """Waits until the given `condition` is true or `timeout` expires.
 
-        The `condition` can be arbitrary JavaScript expression but must contain a 
+        The `condition` can be arbitrary JavaScript expression but must contain a
         return statement (with the value to be returned) at the end.
         See `Execute JavaScript` for information about accessing the
         actual contents of the window through JavaScript.
@@ -113,7 +116,7 @@ class _WaitingKeywords(KeywordGroup):
 
         `error` can be used to override the default error message.
 
-        See also `Wait Until Page Contains`, `Wait Until Page Contains 
+        See also `Wait Until Page Contains`, `Wait Until Page Contains
         Element`, `Wait For Condition` and BuiltIn keyword `Wait Until Keyword
         Succeeds`.
         """
@@ -126,7 +129,7 @@ class _WaitingKeywords(KeywordGroup):
             else:
                 return error or "Element '%s' was not visible in %s" % (locator, self._format_timeout(timeout))
         self._wait_until_no_error(timeout, check_visibility)
-    
+
     def wait_until_element_is_not_visible(self, locator, timeout=None, error=None):
         """Waits until element specified with `locator` is not visible.
 
@@ -136,7 +139,7 @@ class _WaitingKeywords(KeywordGroup):
 
         `error` can be used to override the default error message.
 
-        See also `Wait Until Page Contains`, `Wait Until Page Contains 
+        See also `Wait Until Page Contains`, `Wait Until Page Contains
         Element`, `Wait For Condition` and BuiltIn keyword `Wait Until Keyword
         Succeeds`.
         """
@@ -231,7 +234,7 @@ class _WaitingKeywords(KeywordGroup):
         self._wait_until_no_error(timeout, wait_func)
 
     def _wait_until_no_error(self, timeout, wait_func, *args):
-        timeout = robot.utils.timestr_to_secs(timeout) if timeout is not None else self._timeout_in_secs
+        timeout = timestr_to_secs(timeout) if timeout is not None else self._timeout_in_secs
         maxtime = time.time() + timeout
         while True:
             timeout_error = wait_func(*args)
@@ -241,5 +244,5 @@ class _WaitingKeywords(KeywordGroup):
             time.sleep(0.2)
 
     def _format_timeout(self, timeout):
-        timeout = robot.utils.timestr_to_secs(timeout) if timeout is not None else self._timeout_in_secs
-        return robot.utils.secs_to_timestr(timeout)
+        timeout = timestr_to_secs(timeout) if timeout is not None else self._timeout_in_secs
+        return secs_to_timestr(timeout)