+++ src/Selenium2Library/utils/events/scope_event.py
@@ -1,7 +1,10 @@
-from event import Event
 from robot.libraries.BuiltIn import BuiltIn
 
+from .event import Event
+
+
 class ScopeEvent(Event):
+
     def __init__(self, scope, action, *args, **kwargs):
         self.scope = scope
         self.action = action
@@ -17,8 +20,10 @@ class ScopeEvent(Event):
         if args[0] == self.scope:
             self.action(*self.action_args, **self.action_kwargs)
 
+
 class ScopeStart(ScopeEvent):
     name = 'scope_start'
+
 
 class ScopeEnd(ScopeEvent):
     name = 'scope_end'