+++ test/unit/keywords/test_browsermanagement.py
@@ -1,10 +1,17 @@
 import unittest
-from Selenium2Library.keywords._browsermanagement import _BrowserManagementKeywords
+#Jari 
+## from Selenium2Library.keywords._browsermanagement import _BrowserManagementKeywords
 from Selenium2Library import keywords
-from selenium import webdriver
-from mockito import *
+#Jari 
 
-class BrowserManagementTests(unittest.TestCase): 
+from mockito import when, mock, verify, verifyNoMoreInteractions
+from selenium import webdriver
+
+from Selenium2Library.keywords.browsermanagement import BrowserManagementKeywords
+
+
+class BrowserManagementTests(unittest.TestCase):
+
 
     def test_create_firefox_browser(self):
         test_browsers = ((webdriver.Firefox, "ff"), (webdriver.Firefox, "firEfOx"))
@@ -45,7 +52,7 @@ class BrowserManagementTests(unittest.TestCase):
         self.verify_browser(webdriver.Remote, "htmlunitwithjs")
 
     def test_parse_capabilities_string(self):
-        bm = _BrowserManagementKeywords()
+        bm = BrowserManagementKeywords()
         expected_caps = "key1:val1,key2:val2"
         capabilities = bm._parse_capabilities_string(expected_caps)
         self.assertTrue("val1", capabilities["key1"])
@@ -53,7 +60,7 @@ class BrowserManagementTests(unittest.TestCase):
         self.assertTrue(2, len(capabilities))
 
     def test_parse_complex_capabilities_string(self):
-        bm = _BrowserManagementKeywords()
+        bm = BrowserManagementKeywords()
         expected_caps = "proxyType:manual,httpProxy:IP:port"
         capabilities = bm._parse_capabilities_string(expected_caps)
         self.assertTrue("manual", capabilities["proxyType"])
@@ -77,7 +84,7 @@ class BrowserManagementTests(unittest.TestCase):
         self.assertFalse("some_cap" in webdriver.DesiredCapabilities.CHROME)
 
     def test_set_selenium_timeout_only_affects_open_browsers(self):
-        bm = _BrowserManagementKeywords()
+        bm = BrowserManagementKeywords()
         first_browser, second_browser = mock(), mock()
         bm._cache.register(first_browser)
         bm._cache.close()
@@ -92,12 +99,12 @@ class BrowserManagementTests(unittest.TestCase):
         verifyNoMoreInteractions(second_browser)
 
     def test_bad_browser_name(self):
-        bm = _BrowserManagementKeywords()
+        bm = BrowserManagementKeywords()
         try:
             bm._make_browser("fireox")
             self.fail("Exception not raised")
         except ValueError as e:
-            self.assertEquals("fireox is not a supported browser.", e.message)
+            self.assertEquals(str(e), "fireox is not a supported browser.")
 
     def test_create_webdriver(self):
         bm = _BrowserManagementWithLoggingStubs()
@@ -140,7 +147,7 @@ class BrowserManagementTests(unittest.TestCase):
             def __init__(self, command_executor, sid, **kwargs):
                 FakeReusable.session_id = sid
                 FakeReusable.curl = command_executor
-        keywords._browsermanagement._ReusableDriver = FakeReusable
+        keywords.browsermanagement._ReusableDriver = FakeReusable
         try:
             bm.create_webdriver('FakeWebDriver', 'fake')
             (sid, url, pid) = bm.save_webdriver(None)
@@ -158,7 +165,7 @@ class BrowserManagementTests(unittest.TestCase):
 
     def verify_browser(self , webdriver_type , browser_name, **kw):
         #todo try lambda *x: was_called = true
-        bm = _BrowserManagementKeywords()
+        bm = BrowserManagementKeywords()
         old_init = webdriver_type.__init__
         webdriver_type.__init__ = self.mock_init
 
@@ -175,10 +182,10 @@ class BrowserManagementTests(unittest.TestCase):
         self.was_called = True
 
 
-class _BrowserManagementWithLoggingStubs(_BrowserManagementKeywords):
+class _BrowserManagementWithLoggingStubs(BrowserManagementKeywords):
 
     def __init__(self):
-        _BrowserManagementKeywords.__init__(self)
+        BrowserManagementKeywords.__init__(self)
         def mock_logging_method(self, *args, **kwargs):
             pass
         for name in ['_info', '_debug', '_warn', '_log', '_html']: